+++ packages/plugins/src/plugins/morpho/index.ts
@@ -4,8 +4,8 @@ import { Job } from '../../Job';
 import aaveV3CollateralFetcher from './aaveV3CollateralFetcher';
 import aaveV3SupplyOnlyFetcher from './aaveV3SupplyOnlyFetcher';
 import aaveV3Job from './aaveV3Job';
-import getLendingFetcher from './getLendingFetcher';
-import getYieldFetcher from './getYieldFetcher';
+import { getLendingFetcher } from './lendingFetcher';
+import { getYieldFetcher } from './yieldFetcher';
 import { marketsJob } from './marketsJob';
 import { vaultsJob } from './vaultsJob';
